move orca tf testcase
