+++ python/orca/dev/test/run-pytests-horovod-tf.sh
@@ -27,6 +27,5 @@ cd ../../
 echo "Running RayOnSpark tests"
 python -m pytest -v test/bigdl/orca/learn/ray/tf/ \
     --ignore=test/bigdl/orca/learn/ray/tf/test_ray_tf2estimator.py \
-    --ignore=test/bigdl/orca/learn/ray/tf/test_tf2estimator_ray_dataset.py \
-    --ignore=test/bigdl/orca/learn/ray/tf/test_tf2_estimator_orca_dataset.py
+    --ignore=test/bigdl/orca/learn/ray/tf/test_tf2estimator_ray_dataset.py
 python -m pytest -v test/bigdl/orca/data/test_read_parquet_images.py
